Day7/pyproject.toml
[project]
name = "day7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]
